+++ sofar/__init__.py
@@ -6,12 +6,12 @@ __author__ = """The pyfar developers"""
 __email__ = 'info@pyfar.org'
 __version__ = '0.3.1'
 
+from .sofa import Sofa
 
-from .sofar import (Sofa,
-                    update_conventions,
+from .io import (read_sofa, write_sofa)
+
+from .utils import (update_conventions,
                     list_conventions,
-                    read_sofa,
-                    write_sofa,
                     equals)
